Create VideoDataUCI
To extract the hits you need to include visitid, fullvisitorId, and hitNumber because those are the three IDs that uniquely identify a particular 'hit' (with that then you can extract dimensions). Inner to select data at the individual hit level and an outer to aggregate data over the fields that interest you.